+++ features/cerberusweb.core/api/uri/profiles/calendar.php
@@ -17,9 +17,11 @@
 
 class PageSection_ProfilesCalendar extends Extension_PageSection {
 	function render() {
-		$tpl = DevblocksPlatform::getTemplateService();
+		$tpl = DevblocksPlatform::services()->template();
 		$visit = CerberusApplication::getVisit();
 		$translate = DevblocksPlatform::getTranslationService();
+		
+		$context = CerberusContexts::CONTEXT_CALENDAR;
 		$active_worker = CerberusApplication::getActiveWorker();
 		
 		$response = DevblocksPlatform::getHttpResponse();
@@ -34,6 +36,13 @@ class PageSection_ProfilesCalendar extends Extension_PageSection {
 			return;
 		}
 		$tpl->assign('calendar', $calendar);
+		
+		// Dictionary
+		$labels = array();
+		$values = array();
+		CerberusContexts::getContext($context, $calendar, $labels, $values, '', true, false);
+		$dict = DevblocksDictionaryDelegate::instance($values);
+		$tpl->assign('dict', $dict);
 	
 		// Tab persistence
 		
@@ -64,17 +73,17 @@ class PageSection_ProfilesCalendar extends Extension_PageSection {
 	
 		// Custom Fields
 
-		@$values = array_shift(DAO_CustomFieldValue::getValuesByContextIds(CerberusContexts::CONTEXT_CALENDAR, $calendar->id)) or array();
+		@$values = array_shift(DAO_CustomFieldValue::getValuesByContextIds($context, $calendar->id)) or array();
 		$tpl->assign('custom_field_values', $values);
 		
-		$properties_cfields = Page_Profiles::getProfilePropertiesCustomFields(CerberusContexts::CONTEXT_CALENDAR, $values);
+		$properties_cfields = Page_Profiles::getProfilePropertiesCustomFields($context, $values);
 		
 		if(!empty($properties_cfields))
 			$properties = array_merge($properties, $properties_cfields);
 		
 		// Custom Fieldsets
 
-		$properties_custom_fieldsets = Page_Profiles::getProfilePropertiesCustomFieldsets(CerberusContexts::CONTEXT_CALENDAR, $calendar->id, $values);
+		$properties_custom_fieldsets = Page_Profiles::getProfilePropertiesCustomFieldsets($context, $calendar->id, $values);
 		$tpl->assign('properties_custom_fieldsets', $properties_custom_fieldsets);
 		
 		// Counts
@@ -88,10 +97,10 @@ class PageSection_ProfilesCalendar extends Extension_PageSection {
 		// Link counts
 		
 		$properties_links = array(
-			CerberusContexts::CONTEXT_CALENDAR => array(
+			$context => array(
 				$calendar->id => 
 					DAO_ContextLink::getContextLinkCounts(
-						CerberusContexts::CONTEXT_CALENDAR,
+						$context,
 						$calendar->id,
 						array(CerberusContexts::CONTEXT_CUSTOM_FIELDSET)
 					),
@@ -104,17 +113,14 @@ class PageSection_ProfilesCalendar extends Extension_PageSection {
 		
 		$tpl->assign('properties', $properties);
 			
-		// Macros
-		
-		$macros = DAO_TriggerEvent::getReadableByActor(
-			$active_worker,
-			'event.macro.calendar'
-		);
-		$tpl->assign('macros', $macros);
-
 		// Tabs
-		$tab_manifests = Extension_ContextProfileTab::getExtensions(false, CerberusContexts::CONTEXT_CALENDAR);
+		$tab_manifests = Extension_ContextProfileTab::getExtensions(false, $context);
 		$tpl->assign('tab_manifests', $tab_manifests);
+		
+		// Interactions
+		$interactions = Event_GetInteractionsForWorker::getInteractionsByPointAndWorker('record:' . $context, $dict, $active_worker);
+		$interactions_menu = Event_GetInteractionsForWorker::getInteractionMenu($interactions);
+		$tpl->assign('interactions_menu', $interactions_menu);
 		
 		// Template
 		$tpl->display('devblocks:cerberusweb.core::profiles/calendar.tpl');
@@ -136,6 +142,9 @@ class PageSection_ProfilesCalendar extends Extension_PageSection {
 		try {
 		
 			if(!empty($id) && !empty($do_delete)) { // Delete
+				if(!$active_worker->hasPriv(sprintf("contexts.%s.delete", CerberusContexts::CONTEXT_CALENDAR)))
+					throw new Exception_DevblocksAjaxValidationError(DevblocksPlatform::translate('error.core.no_acl.delete'));
+				
 				DAO_Calendar::delete($id);
 				
 				echo json_encode(array(
@@ -146,9 +155,6 @@ class PageSection_ProfilesCalendar extends Extension_PageSection {
 				return;
 				
 			} else {
-				if(empty($name))
-					throw new Exception_DevblocksAjaxValidationError("The 'Name' field is required.","name");
-				
 				// Owner
 				
 				@list($owner_context, $owner_context_id) = explode(':', $owner);
@@ -167,12 +173,6 @@ class PageSection_ProfilesCalendar extends Extension_PageSection {
 						break;
 				}
 
-				if(empty($owner_context))
-					throw new Exception_DevblocksAjaxValidationError("The 'Owner' field is required.", 'owner');
-				
-				if(!CerberusContexts::isOwnableBy($owner_context, $owner_context_id, $active_worker))
-					throw new Exception_DevblocksAjaxValidationError("You don't have permission to use this owner.", 'owner');
-				
 				// Clean params
 				// [TODO] Move this
 				
@@ -196,8 +196,16 @@ class PageSection_ProfilesCalendar extends Extension_PageSection {
 						DAO_Calendar::PARAMS_JSON => json_encode($params),
 					);
 					
+					if(!DAO_Calendar::validate($fields, $error))
+						throw new Exception_DevblocksAjaxValidationError($error);
+					
+					if(!DAO_Calendar::onBeforeUpdateByActor($active_worker, $fields, null, $error))
+						throw new Exception_DevblocksAjaxValidationError($error);
+					
 					if(false == ($id = DAO_Calendar::create($fields)))
 						return new Exception_DevblocksAjaxValidationError("An unexpected error occurred while saving the record.");
+					
+					DAO_Calendar::onUpdateByActor($active_worker, $fields, $id);
 					
 					if(!empty($view_id) && !empty($id))
 						C4_AbstractView::setMarqueeContextCreated($view_id, CerberusContexts::CONTEXT_CALENDAR, $id);
@@ -216,7 +224,14 @@ class PageSection_ProfilesCalendar extends Extension_PageSection {
 					
 					$change_fields = Cerb_ORMHelper::uniqueFields($fields, $calendar);
 					
+					if(!DAO_Calendar::validate($fields, $error, $id))
+						throw new Exception_DevblocksAjaxValidationError($error);
+					
+					if(!DAO_Calendar::onBeforeUpdateByActor($active_worker, $fields, $id, $error))
+						throw new Exception_DevblocksAjaxValidationError($error);
+					
 					DAO_Calendar::update($id, $change_fields);
+					DAO_Calendar::onUpdateByActor($active_worker, $change_fields, $id);
 				}
 	
 				// Custom fields
